# perf: Avoid `Vec<&str>` allocation during line length checking